+++ packages/guide/src/app/journal/components/notes/notes.component.ts
@@ -1,6 +1,6 @@
 import { Component } from '@angular/core';
 import { BaseTabbedSelectableContainerComponent } from "../../../shared/components/base-tabbed-selectable-container/base-tabbed-selectable-container.component";
-import { MailData } from "@ci/data-types";
+import { MailData, TornPageData } from "@ci/data-types";
 import { combineLatest, Observable, tap } from "rxjs";
 
 @Component({
@@ -8,15 +8,22 @@ import { combineLatest, Observable, tap } from "rxjs";
     templateUrl: './notes.component.html',
 })
 export class NotesComponent extends BaseTabbedSelectableContainerComponent<MailData> {
+    selectedTornPage?: TornPageData;
+    showTornPagesTable = false
     protected mails: MailData[] = [];
-    protected data$: Observable<[mails: MailData[]]>;
-
+    protected tornPages: TornPageData[] = [];
+    protected data$: Observable<[mails: MailData[], tornPages: TornPageData[]]>;
+    protected tabNames = ['Letters', 'Torn pages'];
 
     constructor() {
         super()
-        this.data$ = combineLatest([this._database.fetchMailData$()]).pipe(
-            tap(([mails]) => {
+        this.data$ = combineLatest([
+            this._database.fetchMailData$(),
+            this._database.fetchTornPagesData$()
+        ]).pipe(
+            tap(([mails, tornPages]) => {
                     this.mails = mails;
+                this.tornPages = tornPages;
 
                     const selectedId = this.selectedId;
                     if (selectedId) {
@@ -24,12 +31,30 @@ export class NotesComponent extends BaseTabbedSelectableContainerComponent<MailD
                         if (foundMail) {
                             this.showDetails(foundMail)
                         }
+
+                        const foundTornPage = this.tornPages.find(m => m.key.toLowerCase() === selectedId.toLowerCase())
+                        if (foundTornPage) {
+                            this.showTornPageDetails(foundTornPage)
+                        }
                     }
+
+                this.activateTabFromRoute(this.tabNames)
 
                 }
             )
         );
 
+    }
+
+    override showDetails(selectedEntry?: MailData) {
+        this.selectedTornPage = undefined;
+        super.showDetails(selectedEntry);
+    }
+
+    showTornPageDetails(tornPage: TornPageData): void {
+        this.selectedEntity = undefined;
+        this.selectedTornPage = tornPage;
+        this.openDrawer = true;
     }
 
 }