+++ packages/pak-files-parser/src/main.ts
@@ -45,6 +45,7 @@ import { ItemProcessorMapGenerator } from "./app/item-processor-map.generator";
 import { CookingRecipesChecklistGenerator } from "./app/cooking-recipes-checklist.generator";
 import { CalendarGenerator } from "./app/calendar.generator";
 import { MailDataGenerator } from "./app/mail-data.generator";
+import { TornPagesGenerator } from "./app/torn-pages.generator";
 
 console.log('CURRENT ENVIRONMENT SET TO ' + chalk.bold(environment.isBeta ? 'BETA' : 'LIVE') + '\n');
 
@@ -130,6 +131,8 @@ try {
 
 
         liveGenerators = {
+
+            'torn-pages': new TornPagesGenerator(),
 
             'concerned-monkey-shop-items': {
                 generate: () => new ShopItemDataGenerator(itemDbMap, 'ProjectCoral/Content/ProjectCoral/Core/Data/Shops/DT_ShopConcernedMonke.json').generate({