+++ packages/pak-files-parser/src/types/color.type.ts
@@ -1,0 +1,9 @@
+export type Color = {
+    "R": number
+    "G": number
+    "B": number
+
+
+    "A": number,
+    "Hex": string
+};